apply changes from suggestion
