+++ clients/algoliasearch-client-javascript/packages/client-common/src/types/Host.ts
@@ -3,3 +3,10 @@ export type Host = {
   accept: 'read' | 'readWrite' | 'write';
   protocol: 'http' | 'https';
 };
+
+export type StatefulHost = Host & {
+  status: 'down' | 'timedout' | 'up';
+  lastUpdate: number;
+  isUp: () => boolean;
+  isTimedout: () => boolean;
+};
