+++ clients/algoliasearch-client-javascript/packages/client-common/src/types/index.ts
@@ -1,6 +1,5 @@
 export * from './Cache';
 export * from './CreateClient';
-export * from './Errors';
 export * from './Host';
 export * from './Requester';
 export * from './Transporter';
